+++ packages/widget.vite-plugin.unstable/src/constants.ts
@@ -17,3 +17,4 @@
 export const PALANTIR_PATH: string = ".palantir";
 export const SETUP_PATH: string = `${PALANTIR_PATH}/setup`;
 export const VITE_INJECTIONS: string = `${PALANTIR_PATH}/vite-injections.js`;
+export const CONFIG_FILE_SUFFIX = ".config";
